Fix Clang portability errors in asset tools

+++ lib/TbUiLib/src/mcp/McpAssetTools.cpp
@@ -362,7 +362,9 @@ QJsonObject assetJson(const BrowserAsset& asset)
   if (asset.lastModified)
   {
     json.insert(
-      "lastModified", QString::number(asset.lastModified->time_since_epoch().count()));
+      "lastModified",
+      QString::number(
+        static_cast<qlonglong>(asset.lastModified->time_since_epoch().count())));
   }
 
   return json;
